+++ sdk/dotnet/Network/V20160401/RecordSet.cs
@@ -131,6 +131,8 @@ namespace Pulumi.AzureNative.Network.V20160401
                 Aliases =
                 {
                     new Pulumi.Alias { Type = "azure-nextgen:network/v20160401:RecordSet"},
+                    new Pulumi.Alias { Type = "azure-native:network:RecordSet"},
+                    new Pulumi.Alias { Type = "azure-nextgen:network:RecordSet"},
                     new Pulumi.Alias { Type = "azure-native:network/v20150504preview:RecordSet"},
                     new Pulumi.Alias { Type = "azure-nextgen:network/v20150504preview:RecordSet"},
                     new Pulumi.Alias { Type = "azure-native:network/v20170901:RecordSet"},