+++ sdk/go/azure/network/getRecordSet.go
@@ -7,8 +7,8 @@ import (
 	"github.com/pulumi/pulumi/sdk/v2/go/pulumi"
 )
 
-// Describes a DNS record set (a collection of DNS records with the same name and type) in a Private DNS zone.
-// API Version: 2020-06-01.
+// Describes a DNS record set (a collection of DNS records with the same name and type).
+// API Version: 2018-05-01.
 func LookupRecordSet(ctx *pulumi.Context, args *LookupRecordSetArgs, opts ...pulumi.InvokeOption) (*LookupRecordSetResult, error) {
 	var rv LookupRecordSetResult
 	err := ctx.Invoke("azure-native:network:getRecordSet", args, &rv, opts...)
@@ -19,48 +19,54 @@ func LookupRecordSet(ctx *pulumi.Context, args *LookupRecordSetArgs, opts ...pul
 }
 
 type LookupRecordSetArgs struct {
-	// The name of the Private DNS zone (without a terminating dot).
-	PrivateZoneName string `pulumi:"privateZoneName"`
 	// The type of DNS record in this record set.
 	RecordType string `pulumi:"recordType"`
 	// The name of the record set, relative to the name of the zone.
 	RelativeRecordSetName string `pulumi:"relativeRecordSetName"`
 	// The name of the resource group.
 	ResourceGroupName string `pulumi:"resourceGroupName"`
+	// The name of the DNS zone (without a terminating dot).
+	ZoneName string `pulumi:"zoneName"`
 }
 
-// Describes a DNS record set (a collection of DNS records with the same name and type) in a Private DNS zone.
+// Describes a DNS record set (a collection of DNS records with the same name and type).
 type LookupRecordSetResult struct {
 	// The list of A records in the record set.
 	ARecords []ARecordResponse `pulumi:"aRecords"`
 	// The list of AAAA records in the record set.
 	AaaaRecords []AaaaRecordResponse `pulumi:"aaaaRecords"`
-	// The CNAME record in the record set.
+	// The list of CAA records in the record set.
+	CaaRecords []CaaRecordResponse `pulumi:"caaRecords"`
+	// The CNAME record in the  record set.
 	CnameRecord *CnameRecordResponse `pulumi:"cnameRecord"`
-	// The ETag of the record set.
+	// The etag of the record set.
 	Etag *string `pulumi:"etag"`
 	// Fully qualified domain name of the record set.
 	Fqdn string `pulumi:"fqdn"`
-	// Fully qualified resource Id for the resource. Example - '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Network/privateDnsZones/{privateDnsZoneName}'.
+	// The ID of the record set.
 	Id string `pulumi:"id"`
-	// Is the record set auto-registered in the Private DNS zone through a virtual network link?
-	IsAutoRegistered bool `pulumi:"isAutoRegistered"`
 	// The metadata attached to the record set.
 	Metadata map[string]string `pulumi:"metadata"`
 	// The list of MX records in the record set.
 	MxRecords []MxRecordResponse `pulumi:"mxRecords"`
-	// The name of the resource
+	// The name of the record set.
 	Name string `pulumi:"name"`
+	// The list of NS records in the record set.
+	NsRecords []NsRecordResponse `pulumi:"nsRecords"`
+	// provisioning State of the record set.
+	ProvisioningState string `pulumi:"provisioningState"`
 	// The list of PTR records in the record set.
 	PtrRecords []PtrRecordResponse `pulumi:"ptrRecords"`
 	// The SOA record in the record set.
 	SoaRecord *SoaRecordResponse `pulumi:"soaRecord"`
 	// The list of SRV records in the record set.
 	SrvRecords []SrvRecordResponse `pulumi:"srvRecords"`
+	// A reference to an azure resource from where the dns resource value is taken.
+	TargetResource *SubResourceResponse `pulumi:"targetResource"`
 	// The TTL (time-to-live) of the records in the record set.
 	Ttl *float64 `pulumi:"ttl"`
 	// The list of TXT records in the record set.
 	TxtRecords []TxtRecordResponse `pulumi:"txtRecords"`
-	// The type of the resource. Example - 'Microsoft.Network/privateDnsZones'.
+	// The type of the record set.
 	Type string `pulumi:"type"`
 }