+++ sdk/go/azure/network/v20180301preview/recordSet.go
@@ -70,6 +70,12 @@ func NewRecordSet(ctx *pulumi.Context,
 			Type: pulumi.String("azure-nextgen:network/v20180301preview:RecordSet"),
 		},
 		{
+			Type: pulumi.String("azure-native:network:RecordSet"),
+		},
+		{
+			Type: pulumi.String("azure-nextgen:network:RecordSet"),
+		},
+		{
 			Type: pulumi.String("azure-native:network/v20150504preview:RecordSet"),
 		},
 		{