+++ sdk/go/azure/network/v20180901/init.go
@@ -21,10 +21,10 @@ func (m *module) Version() semver.Version {
 
 func (m *module) Construct(ctx *pulumi.Context, name, typ, urn string) (r pulumi.Resource, err error) {
 	switch typ {
+	case "azure-native:network/v20180901:PrivateRecordSet":
+		r, err = NewPrivateRecordSet(ctx, name, nil, pulumi.URN_(urn))
 	case "azure-native:network/v20180901:PrivateZone":
 		r, err = NewPrivateZone(ctx, name, nil, pulumi.URN_(urn))
-	case "azure-native:network/v20180901:RecordSet":
-		r, err = NewRecordSet(ctx, name, nil, pulumi.URN_(urn))
 	case "azure-native:network/v20180901:VirtualNetworkLink":
 		r, err = NewVirtualNetworkLink(ctx, name, nil, pulumi.URN_(urn))
 	default: