+++ CHANGELOG.md
@@ -9,7 +9,10 @@ CHANGELOG
   from earlier versions, first upgrade to 0.8.0, replace all resources from `latest` modules with their
   counterparts in top-level unnamed modules, apply the change with `pulumi up`, and then upgrade to 0.9.0+.
   [#712](https://github.com/pulumi/pulumi-azure-native/issues/712)
-
+- The ambiguous resource `network.RecordSet` is split into `network.PrivateRecordSet` (private DNS)
+  and `network.RecordSet` (public DNS).
+  [#583](https://github.com/pulumi/pulumi-azure-native/issues/583)
+  
 ---
 
 ## 0.8.0 (2020-04-02)